+++ packages/terra-time-input/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added the changes for inputAttribute aria-describedby.
+
 ## 4.61.0 - (January 25, 2024)
 
 * Changed